Comply with interceptor naming specification
