+++ agent/agent-plugins/httpclient-okhttp32/src/main/java/org/bithon/agent/plugin/httpclient/okhttp32/RealCall$GetResponseWithInterceptorChain.java
@@ -14,13 +14,14 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.httpclient.okhttp3;
+package org.bithon.agent.plugin.httpclient.okhttp32;
 
 import okhttp3.Call;
 import okhttp3.Request;
 import okhttp3.Response;
-import okhttp3.internal.connection.RealConnection;
-import okhttp3.internal.http.RetryAndFollowUpInterceptor;
+import okhttp3.internal.Internal;
+import okhttp3.internal.http.StreamAllocation;
+import okhttp3.internal.io.RealConnection;
 import okio.BufferedSource;
 import org.bithon.agent.bootstrap.aop.AbstractInterceptor;
 import org.bithon.agent.bootstrap.aop.AopContext;
@@ -41,59 +42,51 @@ import java.util.stream.Collectors;
  *
  * @author frankchen
  */
-public class OkHttp3Interceptor extends AbstractInterceptor {
-    private static Logger log = LoggerFactory.getLogger(OkHttp3Interceptor.class);
+public class RealCall$GetResponseWithInterceptorChain extends AbstractInterceptor {
+    private static final Logger log = LoggerFactory.getLogger(RealCall$GetResponseWithInterceptorChain.class);
 
     private HttpOutgoingMetricsCollector metricCollector;
     private Set<String> ignoredSuffixes;
 
     @Override
     public boolean initialize() {
-
         ignoredSuffixes = Arrays.stream("html, js, css, jpg, gif, png, swf, ttf, ico, woff, woff2, json, eot, svg".split(
-                                    ","))
+            ","))
                                 .map(x -> x.trim().toLowerCase(Locale.ENGLISH))
                                 .collect(Collectors.toSet());
 
         metricCollector = MetricCollectorManager.getInstance()
-                                                .getOrRegister("okhttp3", HttpOutgoingMetricsCollector.class);
+                                                .getOrRegister("okhttp3.2", HttpOutgoingMetricsCollector.class);
 
         return true;
     }
 
     @Override
     public InterceptionDecision onMethodEnter(AopContext aopContext) {
-        Call realCall = (Call) aopContext.getTarget();
-        Request request = realCall.request();
-        String uri = request.url().uri().toString().split("\\?")[0];
-
+        Call realCall = aopContext.castTargetAs();
+        Request originRequest = realCall.request();
+        String uri = originRequest.url().uri().toString().split("\\?")[0];
         return needIgnore(uri) ? InterceptionDecision.SKIP_LEAVE : InterceptionDecision.CONTINUE;
     }
 
     @Override
     public void onMethodLeave(AopContext aopContext) {
         Call realCall = (Call) aopContext.getTarget();
-        Request request = realCall.request();
+        Request originRequest = realCall.request();
 
-        String uri = request.url().uri().toString();
-        String httpMethod = request.method().toUpperCase(Locale.ENGLISH);
+        String requestUri = originRequest.url().uri().toString();
+        String requestMethod = originRequest.method().toUpperCase(Locale.ENGLISH);
 
         if (aopContext.getException() != null) {
-            metricCollector.addExceptionRequest(uri,
-                                                httpMethod,
+            metricCollector.addExceptionRequest(requestUri,
+                                                requestMethod,
                                                 aopContext.getCostTime());
         } else {
             Response response = aopContext.castReturningAs();
-            metricCollector.addRequest(uri,
-                                       httpMethod,
-                                       response.code(),
-                                       aopContext.getCostTime());
+            metricCollector.addRequest(requestUri, requestMethod, response.code(), aopContext.getCostTime());
         }
 
-        this.addBytes(uri,
-                      httpMethod,
-                      request,
-                      realCall);
+        addBytes(requestUri, requestMethod, originRequest, realCall);
     }
 
     private boolean needIgnore(String uri) {
@@ -101,44 +94,26 @@ public class OkHttp3Interceptor extends AbstractInterceptor {
         return ignoredSuffixes.contains(suffix);
     }
 
-    private void addBytes(String uri,
-                          String httpMethod,
+    private void addBytes(String uri, String method,
                           Request request,
                           Call call) {
         try {
-            log.debug("okhttp 3.4 getByteSize");
-
             long requestByteSize = 0;
             long responseByteSize = 0;
             if (request.body() != null) {
                 requestByteSize = request.body().contentLength() < 0 ? 0 : request.body().contentLength();
             }
 
-            RealConnection realConnection = null;
-            // 3.4
-            RetryAndFollowUpInterceptor retryAndFollowUpInterceptor = (RetryAndFollowUpInterceptor) ReflectionUtils.getFieldValue(
-                call,
-                "retryAndFollowUpInterceptor");
-            if (retryAndFollowUpInterceptor != null) {
-                realConnection = retryAndFollowUpInterceptor.streamAllocation().connection();
-            } else {
-                // 3.2.0
-                // Object HttpEngine = ReflectionUtils.getFieldValue(call, "engine");
-                // StreamAllocation streamAllocation = (StreamAllocation)
-                // ReflectionUtils.getFieldValue(HttpEngine, "streamAllocation");
-                // realConnection = streamAllocation.connection();
-            }
-
+            StreamAllocation streamAllocation = Internal.instance.callEngineGetStreamAllocation(call);
+            RealConnection realConnection = streamAllocation.connection();
             if (realConnection != null) {
                 responseByteSize = ((BufferedSource) ReflectionUtils.getFieldValue(realConnection, "source")).buffer()
                                                                                                              .size();
             }
 
-            log.debug("OKHttp ,request requestByteSize {},responseByteSize {}", requestByteSize, responseByteSize);
-
-            metricCollector.addBytes(uri, httpMethod, requestByteSize, responseByteSize);
+            metricCollector.addBytes(uri, method, requestByteSize, responseByteSize);
         } catch (Exception e) {
-            log.error("OKHttp getByteSize", e);
+            log.warn("OKHttp getByteSize", e);
         }
     }
 }
